requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.7.2
azure-functions-durable==1.1.3
pytest==7.1.2
pytest-cov==3.0.0
coverage==6.4.0
pytest-mock==3.7.0
pytest-asyncio==0.18.3
